Time: 6 ms (2.04%), Space: 25.5 MB (5.13%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.cpp
@@ -19,11 +19,11 @@ public:
         
         GetOrder(root,0);
         for(int i = 0;i < maxDepth+1;i++){
-            long long sum = 0;
+            double sum = 0;
             for(int j = 0; j < record[i].size() ; j++){
-                sum += record[i][j];
+                sum += double(record[i][j]);
             }
-            double average = double(sum) / double(record[i].size());
+            double average = sum / record[i].size();
             finalAnswer.push_back(average);
         }
